bounding_box fix
